refactor(sdk): move refresh_protocol_version into its own module
Relocate `Sdk::refresh_protocol_version` and its private helper
`extract_network_protocol_version` out of the oversized `src/sdk.rs`
into a dedicated `src/sdk/refresh.rs` submodule. The shared
`min_protocol_version` / `maybe_update_protocol_version` helpers stay in
the parent module; the submodule reaches them via `super::` / `self`.
Tests are unchanged except the two `extract_network_protocol_version`
call sites, now `super::refresh::extract_network_protocol_version`.
